+++ src/dependency_manager.py
@@ -10,7 +10,6 @@ from urllib.error import URLError, HTTPError
 import json
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 
-from utils import run_process
 from loguru import logger
